+++ e2e/utils/stop-local-registry.ts
@@ -1,7 +1,0 @@
-export default () => {
-  // @ts-expect-error: assigned globally by start-and-publish-to-local-registry
-  if (global.stopLocalRegistry) {
-    // @ts-expect-error: assigned globally by start-and-publish-to-local-registry
-    global.stopLocalRegistry();
-  }
-};