+++ packages/builder/jest.config.ts
@@ -1,7 +1,0 @@
-export default {
-  displayName: 'builder',
-  preset: '../../jest.preset.js',
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/packages/builder',
-  coverageReporters: ['text-summary', 'lcov'],
-};